fix: rename path login → auth
